+++ R/matrix.R
@@ -128,8 +128,17 @@ PDfunc <-function(M,func=function(m){1/m},force=FALSE,pseudo=FALSE,tol=.Machine$
   else if(DIM==2)
   {
     TR <- (M[1,1] + M[2,2])/2 # half trace
-    DET <- M[1,1]*M[2,2] - M[1,2]*M[2,1]
-    DET <- TR^2 - DET
+    BIGNUM <- TR^2 > .Machine$double.xmax * .Machine$double.eps
+    if(BIGNUM)
+    {
+      DET <- (M[1,1]/TR)*(M[2,2]/TR) - (M[1,2]/TR)*(M[2,1]/TR)
+      DET <- 1 - DET
+    }
+    else
+    {
+      DET <- M[1,1]*M[2,2] - M[1,2]*M[2,1]
+      DET <- TR^2 - DET
+    }
     if(DET<0) # this shouldn't ever happen with Hermitian matrices
     {
       if(!force && !pseudo) { stop("Matrix not positive definite.") }
@@ -144,6 +153,7 @@ PDfunc <-function(M,func=function(m){1/m},force=FALSE,pseudo=FALSE,tol=.Machine$
     else
     {
       DET <- sqrt(DET) # now root term
+      if(BIGNUM) { DET <- DET * TR }
 
       # hermitian formula
       V <- diag(1/2,2) %o% c(1,1) + ((M-TR*diag(2))/(DET*2)) %o% c(1,-1)
@@ -216,7 +226,7 @@ PDsolve <- function(M,force=FALSE,pseudo=FALSE,tol=.Machine$double.eps)
 
     # regular inverse of remaining dimensions
     REM <- !(INF|ZERO)
-    if(any(REM)) { M[REM,REM] <- PDsolve(M[REM,REM]) }
+    if(any(REM)) { M[REM,REM] <- PDsolve(M[REM,REM,drop=FALSE]) }
 
     return(M)
   }
@@ -351,13 +361,30 @@ conditionNumber <- function(M)
 # Positive definite part of matrix
 PDclamp <- function(M)
 {
-  # symmetrize
-  M <- He(M)
+  # check for Inf
+  INF <- diag(M)==Inf
+  if(any(INF))
+  {
+    if(any(INF))
+    {
+      M[INF,INF] <- 0
+      diag(M)[INF] <- Inf
+    }
 
-  M <- PDfunc(M,function(m){clamp(m,0,Inf)},pseudo=TRUE)
+    # regular clamp of remaining dimensions
+    REM <- !INF
+    if(any(REM)) { M[REM,REM] <- PDclamp(M[REM,REM,drop=FALSE]) }
+  }
+  else
+  {
+    # symmetrize
+    M <- He(M)
 
-  # symmetrize
-  M <- He(M)
+    M <- PDfunc(M,function(m){clamp(m,0,Inf)},pseudo=TRUE)
+
+    # symmetrize
+    M <- He(M)
+  }
 
   return(M)
 }